fix: change sales.item type from float to integer

+++ app/models/sales_model.py
@@ -10,7 +10,7 @@ class SalesModel(db.Model):
     id: int
     seller: str
     buyer: str
-    item: float
+    item: int
     value: float
 
     __tablename__ = "sales"
@@ -18,7 +18,7 @@ class SalesModel(db.Model):
     id = Column(Integer, primary_key=True)
     seller = Column(String(50), ForeignKey("users.user_name"), nullable=False)
     buyer = Column(String(50), ForeignKey("users.user_name"), nullable=False)
-    item = Column(Numeric, ForeignKey("nfts.id"), nullable=False)
+    item = Column(Integer, ForeignKey("nfts.id"), nullable=False)
     value = Column(Numeric, nullable=False)
 
     nfts = relationship("NftsModel", backref="nfts")
